Use ESPHome UART bus instead of initializing a new one
It uses the D1 and D2 (hardware serial 0) for communication with the Roomba.
The logger is set to log to hardware serial 1 to make sure it does not interfere with the roomba communication.
I don't use the distance sensor, so I removed it, but it could easily be added back.